Make mavenOutputChannel static.
