+++ src/mavenOutputChannel.ts
@@ -1,0 +1,31 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+import * as vscode from "vscode";
+
+class MavenOutputChannel {
+    private readonly channel: vscode.OutputChannel = vscode.window.createOutputChannel("Maven for Java");
+
+    public appendLine(message: any, title?: string): void {
+        if (title) {
+            const simplifiedTime: string = (new Date()).toISOString().replace(/z|t/gi, " ").trim(); // YYYY-MM-DD HH:mm:ss.sss
+            const hightlightingTitle: string = `[${title} ${simplifiedTime}]`;
+            this.channel.appendLine(hightlightingTitle);
+        }
+        this.channel.appendLine(message);
+    }
+
+    public append(message: any): void {
+        this.channel.append(message);
+    }
+
+    public show(): void {
+        this.channel.show();
+    }
+
+    public dispose(): void {
+        this.channel.dispose();
+    }
+}
+
+export const mavenOutputChannel: MavenOutputChannel = new MavenOutputChannel();
